Refactor Firestore migration command to use new preview template structure and add document preview functions
